docs: formatDage 함수에 주석 추가
