--- src/utils/dateFunc.js
@@ -1,10 +1,0 @@
-// src/utils/date.js
-export const dateFunc = (createdAt) => {
-  const date = new Date(createdAt);
-
-  const year = date.getFullYear();
-  const month = String(date.getMonth() + 1).padStart(2, "0");
-  const day = String(date.getDate()).padStart(2, "0");
-
-  return `${year}.${month}.${day}`;
-};
